# api_contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "交通法规知识库服务",
    "description": "分层 RAG 管道 + 强制引用式生成，HTTP 封装",
    "version": "0.1.0"
  },
  "paths": {
    "/documents": {
      "post": {
        "summary": "Upload",
        "operationId": "upload_documents_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Upload Documents Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Documents",
        "operationId": "documents_documents_get",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Documents Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{doc_id}": {
      "delete": {
        "summary": "Remove",
        "operationId": "remove_documents__doc_id__delete",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Remove Documents  Doc Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reindex": {
      "post": {
        "summary": "Reindex",
        "operationId": "reindex_reindex_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Reindex Reindex Post"
                }
              }
            }
          }
        }
      }
    },
    "/laws": {
      "get": {
        "summary": "Laws",
        "operationId": "laws_laws_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Laws Laws Get"
                }
              }
            }
          }
        }
      }
    },
    "/articles/lookup": {
      "post": {
        "summary": "Article Lookup",
        "operationId": "article_lookup_articles_lookup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleLookup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Article Lookup Articles Lookup Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/materials/search": {
      "post": {
        "summary": "Materials Search",
        "operationId": "materials_search_materials_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialSearch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Materials Search Materials Search Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/qa": {
      "post": {
        "summary": "Ask",
        "operationId": "ask_qa_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ask Qa Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/answer": {
      "post": {
        "summary": "Generate",
        "operationId": "generate_answer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Generate Answer Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/answer/stream": {
      "post": {
        "summary": "Generate Stream",
        "operationId": "generate_stream_answer_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/qa/stream": {
      "post": {
        "summary": "Ask Stream",
        "operationId": "ask_stream_qa_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnswerRequest": {
        "properties": {
          "question": {
            "$ref": "#/components/schemas/QuestionBody",
            "description": "问题与历史；没有 top_k，生成用不到它"
          },
          "retrieval": {
            "additionalProperties": true,
            "type": "object",
            "title": "Retrieval",
            "description": "已经检索好的 RetrievalResult.to_dict()（POST /qa?mode=search 的返回）"
          },
          "timeliness": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Timeliness",
            "description": "[时效N] 联网片段"
          },
          "materials": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Materials",
            "description": "[材料N] 会话材料片段"
          }
        },
        "type": "object",
        "required": [
          "question",
          "retrieval"
        ],
        "title": "AnswerRequest"
      },
      "ArticleLookup": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "一句话，条号写在里面（与 article_no 二选一，给了 article_no 就按它走）"
          },
          "article_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Article No",
            "description": "条号，中英文数字都行，例如「第九十条」或「90」"
          },
          "law_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Law Name",
            "description": "法规名，可选；多部法规都有这个条号时必须给"
          }
        },
        "type": "object",
        "title": "ArticleLookup"
      },
      "Body_upload_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "docx / md / txt / pdf"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "session=仅本次会话；permanent=入知识库",
            "default": "session"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_documents_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MaterialSearch": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "检索词，用材料里真出现过的说法"
          },
          "doc_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Doc Ids",
            "description": "会话材料 doc_id（POST /documents mode=session 的返回）"
          },
          "top_k": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Top K",
            "description": "返回段数，默认 5",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "MaterialSearch"
      },
      "QaRequest": {
        "properties": {
          "question": {
            "type": "string",
            "minLength": 1,
            "title": "Question",
            "description": "用户问题"
          },
          "mode": {
            "type": "string",
            "enum": [
              "ask",
              "search"
            ],
            "title": "Mode",
            "description": "ask=检索+生成（默认）；search=只检索不花钱",
            "default": "ask"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K",
            "description": "覆盖默认召回条数（默认 6）"
          },
          "pool": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pool",
            "description": "候选池深度，同时按它放行超过 top_k 上限的返回条数（默认取 RAG_CANDIDATES=20，上限 RAG_POOL_MAX）。与 eval 的 --pool 同口径；给了它就盖过 top_k"
          },
          "debug": {
            "type": "boolean",
            "title": "Debug",
            "description": "检索诊断：带上两路通道各自的名次与原始分（该通道没跑时为 null）",
            "default": false
          },
          "law_filter": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Law Filter",
            "description": "只在给定 law_id 内检索（GET /laws 的 law_id）"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "QaRequest"
      },
      "QuestionBody": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text",
            "description": "用户问题"
          },
          "history": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "History",
            "description": "多轮历史 (role, content)，原样进生成提示词"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "QuestionBody"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}